Create pairwise.DA.test.R
The first version of the pairwise differential abundance function. 
It was made to feel similar to the pairwise_DA_wrapper function, but is a lot quicker and takes a clr-transformed count table rather than transforming within the function as its predecessor did. 

Effect sizes are calculated by Cohen's D with Hedge's correction as by the effsize library