Create pairwise_glm.R
The intent of this function is to fit a glm over each feature of a data.frame. 
The model argument should look just like a normal R model, for instance like this:  model = " ~ Age + Group "
Independent variables should be found as column names of your metadata for this to function properly.