Create pairwise_glmer
The intent of this function is to fit a lm with mixed effects over each feature of a data.frame. The function requires lme4 and lmerTest in order to function properly. 
The model argument should look just like a normal lme4 model, for instance like this:  model = " ~ Age + Group + (1|ID)"  note the addition of the random effect. 
Independent variables should be found as column names of your metadata for this to function properly.